eventcalendarbuilder/requirements.txt
annotated-types==0.6.0
arrow==1.3.0
Babel==2.14.0
binaryornot==0.4.4
blinker==1.7.0
blis==0.7.11
briefcase==0.3.17
build==1.0.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
clr-loader==0.2.6
colorama==0.4.6
confection==0.1.4
cookiecutter==2.5.0
cymem==2.0.8
darkdetect==0.8.0
filelock==3.13.1
Flask==3.0.2
fsspec==2024.2.0
gevent==23.9.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
httplib2==0.22.0
huggingface-hub==0.20.3
icalendar==5.0.11
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
preshed==3.0.9
protobuf==4.25.2
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytest==8.0.0
python-dateutil==2.8.2
python-slugify==8.0.4
pythonnet==3.0.3
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.2
setuptools==69.0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.4
srsly==2.4.8
sympy==1.12
text-unidecode==1.3
thinc==8.2.3
toga==0.4.2
toga-core==0.4.2
toga-winforms==0.4.2
tokenizers==0.15.1
tomli_w==1.0.0
torch==2.2.0
tqdm==4.66.1
transformers==4.36.2
travertino==0.3.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wheel==0.42.0
zope.event==5.0
zope.interface==6.1
